requirements.txt
flask
pyrebase
firebase_admin
google.cloud
firebase-admin 
google-cloud-firestore